// src/enums/todo.enum.ts
export enum ConstantsEnum {
    FAKE_TIME_REQUEST = 2000,
    KEY_LOCAL_STORAGE = 'todos'
}

export enum StatusEnum {
    PENDING = 'pendentes',
    COMPLETED = 'concluídas'
}

export enum ActionTypesEnum {
    ADD_TODO = 'addTodo',
    SET_LOADING = 'setLoading',
    GET_ALL_TODOS = 'getAllTodos',
}